Separate into base and domain layers
