+++ .github/workflows/flake.yml
@@ -20,9 +20,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --show-source --statistics --max-complexity=10
+        flake8 . --count --show-source --statistics
