+++ discretesampling/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["discrete", "spectrum","decision_tree"]  # noqa
+__all__ = ["base", "domain"]  # noqa
