+++ discretesampling/base/algorithms/mcmc.py
